add is_upgrade
this is hacky and might fail on different ABIs.
the better fix is to have a C shim to handle this,
but that puts an unnecessary call in the hot path
and more importantly, complicates the build process.
I should work on this later, but I don't care that much.
I have no personal need to run this on anything other than x86_64.